+++ src/com/intellij/idea/plugin/hybris/codeInspection/rule/impex/ImpexUnknownAttributeModifierInspection.kt
@@ -1,6 +1,6 @@
 /*
- * This file is part of "SAP Commerce Developers Toolset" plugin for Intellij IDEA.
- * Copyright (C) 2019-2023 EPAM Systems <hybrisideaplugin@epam.com> and contributors
+ * This file is part of "SAP Commerce Developers Toolset" plugin for IntelliJ IDEA.
+ * Copyright (C) 2019-2024 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -44,7 +44,7 @@ class ImpexUnknownAttributeModifierInspection : LocalInspectionTool() {
                 .flatMap { it.attributeList }
                 .map { it.anyAttributeName }
                 .none { it.textMatches(AttributeModifier.TRANSLATOR.modifierName) }
-            if (AttributeModifier.getByModifierName(attribute.text) == null && noTranslator) {
+            if (AttributeModifier.getModifier(attribute.text) == null && noTranslator) {
                 problemsHolder.registerProblem(
                     attribute,
                     HybrisI18NBundleUtils.message("hybris.inspections.impex.ImpexUnknownAttributeModifierInspection.key", attribute.text),